+++ mandelbrot_standard.h
@@ -3,6 +3,6 @@
 #ifndef MANDELBROT_STANDARD_H
 #define MANDELBROT_STANDARD_H
 
-void calculate_mandelbrot_standard(Zoom zoom, uint16_t max_iterations, uint16_t *result, int thread_count, int lines_per_thread);
+void calculate_mandelbrot_standard(Zoom zoom, uint16_t max_iterations, float *result, int thread_count, int lines_per_thread);
 
 #endif
